third_party/blink/web_tests/external/wpt/css/cssom/cssstyledeclaration-custom-properties-expected.txt

This is a testharness.js-based test.
FAIL Custom properties are included in computed style assert_equals: Should show up in .length expected 343 but got 342
Harness: the test ran to completion.
